# docs/api/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "BASE Wealth Management API",
    "description": "REST API construida con Kotlin y Spring Boot 3 para gestión de activos patrimoniales y proyecciones financieras.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/v1/platforms": {
      "get": {
        "tags": [
          "Platforms"
        ],
        "summary": "Lista todas las plataformas del usuario, incluidas las que no tienen posiciones",
        "operationId": "getAllPlatforms",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlatformResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Platforms"
        ],
        "summary": "Crea una nueva plataforma",
        "operationId": "createPlatform",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlatformRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/holdings": {
      "get": {
        "tags": [
          "Holdings"
        ],
        "summary": "Lista las posiciones del usuario, opcionalmente filtradas por clase o plataforma",
        "operationId": "getAllHoldings",
        "parameters": [
          {
            "name": "assetClass",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "platform",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HoldingResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Holdings"
        ],
        "summary": "Crea una nueva posición de inversión",
        "operationId": "createHolding",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHoldingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HoldingResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/platforms/{name}": {
      "delete": {
        "tags": [
          "Platforms"
        ],
        "summary": "Elimina una plataforma sin posiciones asociadas",
        "operationId": "deletePlatform",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "409": {
            "description": "La plataforma todavía tiene posiciones asociadas"
          }
        }
      },
      "patch": {
        "tags": [
          "Platforms"
        ],
        "summary": "Renombra o recategoriza una plataforma existente",
        "operationId": "patchPlatform",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchPlatformRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/holdings/{id}": {
      "get": {
        "tags": [
          "Holdings"
        ],
        "summary": "Obtiene el detalle de una posición por su ID",
        "operationId": "getHoldingById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HoldingResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Holdings"
        ],
        "summary": "Elimina una posición financiera",
        "operationId": "deleteHolding",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      },
      "patch": {
        "tags": [
          "Holdings"
        ],
        "summary": "Actualiza parcialmente los datos de una posición existente",
        "operationId": "updateHolding",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateHoldingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HoldingResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreatePlatformRequest": {
        "required": [
          "name",
          "type"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        }
      },
      "PlatformResponse": {
        "required": [
          "createdAt",
          "name",
          "type"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateHoldingRequest": {
        "required": [
          "assetClass",
          "name",
          "platform",
          "valueUsd"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "assetClass": {
            "type": "string"
          },
          "platform": {
            "type": "string"
          },
          "valueUsd": {
            "minimum": 0.0,
            "exclusiveMinimum": false,
            "type": "number",
            "format": "double"
          }
        }
      },
      "HoldingResponse": {
        "required": [
          "assetClass",
          "createdAt",
          "id",
          "name",
          "platform",
          "updatedAt",
          "valueUsd"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "assetClass": {
            "type": "string"
          },
          "platform": {
            "type": "string"
          },
          "valueUsd": {
            "type": "number",
            "format": "double"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "PatchPlatformRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        }
      },
      "UpdateHoldingRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "assetClass": {
            "type": "string"
          },
          "platform": {
            "type": "string"
          },
          "valueUsd": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
    }
  }
}